2.17*: Phonebook step 12
Exercise description:

"Open your application in two browsers. If you delete a person in browser 1 a short while before attempting to change the person's phone number in browser 2, you will get the following error messages:
---
Fix the issue according to the example shown in promise and errors in part 2. Modify the example so that the user is shown a message when the operation does not succeed. The messages shown for successful and unsuccessful events should look different:
---
Note that even if you handle the exception, the first "404" error message is still printed to the console. But you should not see "Uncaught (in promise) Error"."

-I wrote one hardcoded data (Mr. I don't exist) for easier testing instead of two browser operation, but works also as described in the exercise.